+++ src/services/__tests__/shippo-fulfillment.js
@@ -215,11 +215,11 @@ describe("ShippoFulfillmentService", () => {
       ).toBe(true)
     })
 
-    it("returns true when supports_return_labels", async () => {
-      expect(
-        await shippoFulfilService.canCalculate({ supports_return_labels: true })
-      ).toBe(true)
-    })
+    // it("returns true when supports_return_labels", async () => {
+    //   expect(
+    //     await shippoFulfilService.canCalculate({ supports_return_labels: true })
+    //   ).toBe(true)
+    // })
 
     it("returns false when free", async () => {
       expect(await shippoFulfilService.canCalculate({ type: "FREE" })).toBe(
@@ -285,10 +285,15 @@ describe("ShippoFulfillmentService", () => {
       jest.clearAllMocks()
     })
 
+    const eventBusService = {
+      emit: jest.fn(),
+    }
+
     const shippoClientService = new ShippoClientService({}, {})
     const shippoFulfilService = new ShippoFulfillmentService({
       shippoClientService,
       totalsService,
+      eventBusService,
     })
 
     const methodData = {
@@ -306,12 +311,14 @@ describe("ShippoFulfillmentService", () => {
     fromOrder.shipping_methods = [
       { shipping_option: mockShippingOption({ variant: "default" }) },
     ]
+    const fulfillment = { id: "123" }
 
     it("returned an object with shippo_order_id prop", async () => {
       const result = await shippoFulfilService.createFulfillment(
         methodData,
         fulfillmentItems,
-        fromOrder
+        fromOrder,
+        fulfillment
       )
       expect(result).toBeObject()
       expect(result).toContainKey("shippo_order_id")
@@ -352,21 +359,21 @@ describe("ShippoFulfillmentService", () => {
       creatReturn
     
     *****************************/
-  // describe("createReturn", () => {
-  //   beforeAll(async () => {
-  //     jest.clearAllMocks()
-  //   })
+  describe("createReturn", () => {
+    beforeAll(async () => {
+      jest.clearAllMocks()
+    })
 
-  //   const shippoClientService = new ShippoClientService({}, {})
-  //   const shippoFulfilService = new ShippoFulfillmentService({
-  //     shippoClientService,
-  //   })
+    const shippoClientService = new ShippoClientService({}, {})
+    const shippoFulfilService = new ShippoFulfillmentService({
+      shippoClientService,
+    })
 
-  //   it("returns resolved promise", async () => {
-  //     expect.assertions(1)
-  //     await expect(shippoFulfilService.createReturn()).resolves.toEqual({})
-  //   })
-  // })
+    it("returns resolved promise", async () => {
+      expect.assertions(1)
+      await expect(shippoFulfilService.createReturn()).resolves.toEqual({})
+    })
+  })
 
   /** **************************